backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
python-dotenv
google-genai
langgraph
langchain-core
requests